long

+++ src/main/java/ru/netology/stats/Java7/services/StatsService.java
@@ -3,71 +3,71 @@ package ru.netology.stats.Java7.services;
 //номер месяца, в котором был минимум продаж, то есть осуществлены продажи на минимальную сумму
 public class StatsService {
     public int minSales(long[] sales) {
-        int minMonth = 0;
+        long minMonth = 0;
 
         for (int i = 0; i < sales.length; i++) {
-            if (sales[i] <= sales[minMonth]) {
+            if (sales[i] <= sales[(int) minMonth]) {
                 minMonth = i;
             }
         }
-        return minMonth + 1;
+        return (int) (minMonth + 1);
     }
 
     // номер месяца, в котором был пик продаж, то есть осуществлены продажи на максимальную сумму
     public int maxSales(long[] sales) {
-        int maxMonth = 0;
+        long maxMonth = 0;
 
         for (int i = 0; i < sales.length; i++) {
-            if (sales[i] >= sales[maxMonth]) {
+            if (sales[i] >= sales[(int) maxMonth]) {
                 maxMonth = i;
             }
         }
-        return maxMonth + 1;
+        return (int) maxMonth + 1;
     }
 
     //сумма всех продаж
     public int sumSales(long[] sales) {
-        int sum = 0;
+        long sum = 0;
 
         for (int i = 0; i < sales.length; i++) {
             sum += (int) sales[i];
         }
-        return sum;
+        return (int) sum;
     }
 
     //средняя сумма продаж в месяц
     public int averageSales(long[] sales) {
-        int sum = 0;
+        long sum = 0;
 
         for (int i = 0; i < sales.length; i++) {
             sum += (int) sales[i];
         }
-        return sum / sales.length;
+        return (int) (sum / sales.length);
     }
 
     //количество месяцев, в которых продажи были выше среднего
     public int moreAverageSales(long[] sales) {
-        int averageSales = averageSales(sales);
-        int moreSales = 0;
+        long averageSales = averageSales(sales);
+        long moreSales = 0;
 
         for (int i = 0; i < sales.length; i++) {
             if (averageSales < sales[i]) {
                 moreSales ++;
             }
         }
-        return moreSales;
+        return (int) moreSales;
     }
 
     //количество месяцев, в которых продажи были ниже среднего
     public int lessAverageSales(long[] sales) {
-        int averageSales = averageSales(sales);
-        int lessSales = 0;
+        long averageSales = averageSales(sales);
+        long lessSales = 0;
 
         for (int i = 0; i < sales.length; i++) {
             if (averageSales > sales[i]) {
                 lessSales ++;
             }
         }
-        return lessSales;
+        return (int) lessSales;
     }
 }
